Fix presentation issue
Adding the view of the presenting view controller to the transition container view causes
an accessibility issue when using the keyboard option. Typical iOS behavior is that a
view controller that was presented on top of another view controller will take all the keyboard
gestures and that the view under will not be accessible with the keyboard.

The issue here is that by adding the presenting view to the container iOS now will make it on the same level
as the presented view with respect to accessibility.

In order to support the change we remove the presenting view from the container, this poses an issue
because the presentation style contains a background which now doesn't make much sense to have as it
is rendered above the presenting view, so we also remove the option to set a custom background color.